Fix: OlFeature do not create a new Feature on props.properties change
Creating a new Feature results in removing any styling that has been applied to the feature.
This can also be solved by adding more reactivity on the injected styledObj in OlStyle.

Please make sure that I do not break anything by removing the usage of `usePropsAsObjectProperties`. I did not figure out what it adds of value in this component. Is it possible to pass in the coordinates of the feature into the props.properties object? -> If so, then I break the functionality here

See https://github.com/MelihAltintas/vue3-openlayers/issues/128#issuecomment-1630472490 for details

Fixes #128