perf: prevent watching reactive props continously
BREAKING CHANGE:
the composable `usePropsAsObjectProperties` which is used as foundation in all components has been refactored.
Props were watched always deeply and for lot's of components this was done multiple times since the components implement watchers by themselves.
However, even if not expected, this may cause some side effects with reactive changes which must be applied for individual components when not implemented and therefore considered as a breaking change.
Please check your app carefully when updating to the new major version.