refactor(ol-layer-*):  push layer to correct collection based on parent component injection
BREAKING CHANGE:
These changes may not affect you directly. Previously, for some layer types, a layer was always added to the map, even if the layer component was used within a LayerGroup. This resulted in layers existing multiple times (in the LayerGroup and at the top level (Reported in #344)). Due to the refactoring, all layers are now treated the same (with onMounted and when changing layer properties).
1.) If the layer component is within an `ol-layer-group`, it is added to the group.
2.) If the layer component is inside an `ol-overviewmap-control`, it is added to the OverviewMap
3.) If 1) and 2) do not apply, the layer component is added directly to the map

partially fixes #344