fix(ol-swipe-control): explicitly set `layers` and `rightLayers`
Before, only the second passed layer was added to the right side. Now you can explicitly pass (multiple) layers for the left (or top) using `layers` and for the right (bottom) using `rightLayers`.

BREAKING CHANGE: `ol-swipe-control`: property `layerList` has been removed and splitted into `layers` and `rightLayers`. Please pass explicitly `layers` for the left side (or top, depending on the `orientation`) and `rightLayers` for the right side (or bottom, depending on the `orientation`)
Closes: #376